refactor: layout fsd 아키텍처 기반으로 구조 변경
