+++ src/app/(app)/layout/components/CategoryIcon.tsx
@@ -1,0 +1,13 @@
+import { Icons } from '@/shared/ui';
+
+import { iconMap } from '../lib/iconMapping';
+
+interface CategoryIconProps {
+  categoryName: string;
+  className?: string;
+}
+
+export function CategoryIcon({ categoryName, className = 'h-4 w-4' }: CategoryIconProps) {
+  const IconComponent = Icons[iconMap[categoryName] ?? 'circle'];
+  return <IconComponent className={className} />;
+}
