+++ src/app/(app)/layout/hooks/useCategories.ts
@@ -1,0 +1,27 @@
+import { useEffect, useState } from 'react';
+
+import { CategoryGroup, getCategories } from '../api/categories';
+
+export const useCategories = () => {
+  const [categories, setCategories] = useState<CategoryGroup[]>([]);
+  const [isLoading, setIsLoading] = useState(true);
+  const [error, setError] = useState<Error | null>(null);
+
+  useEffect(() => {
+    const fetchCategories = async () => {
+      try {
+        const response = await getCategories();
+        setCategories(response.data.categories);
+      } catch (err) {
+        console.error('카테고리 로딩 실패:', err);
+        setError(err as Error);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchCategories();
+  }, []);
+
+  return { categories, isLoading, error };
+};
