+++ src/app/(app)/layout/index.ts
@@ -1,0 +1,2 @@
+export { Header } from './ui/Header';
+export { Sidebar } from './ui/Sidebar';
