+++ src/app/(app)/layout/lib/iconMapping.ts
@@ -1,0 +1,16 @@
+import { Icons } from '@/shared/ui';
+
+// 순수 매핑 데이터
+export const iconMap: Record<string, keyof typeof Icons> = {
+  취업: 'building2',
+  연애: 'heart',
+  경제: 'wallet',
+  정치: 'landPlot',
+  스포츠: 'dumbbell',
+  사회: 'users',
+  익명자유: 'smile',
+  익명고민: 'helpCircle',
+  반려동물: 'dog',
+  무한위로: 'handshake',
+  응원합시다: 'users',
+};
